% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate_recipe_steps.R
\name{activate_recipe_step}
\alias{activate_recipe_step}
\title{Activate (i.e., set active = TRUE) for a given recipe step.}
\usage{
activate_recipe_step(connection, recipe_uid, step_uid)
}
\arguments{
\item{connection}{A connection object, as retrieved from \code{\link[=connect_scrapebot]{connect_scrapebot()}}.}

\item{recipe_uid}{The \code{uid} (integer) of the corresponding recipe.}

\item{step_uid}{The \code{uid} (integer) of the corresponding recipe step.}
}
\value{
\code{TRUE} if update was successful, \code{FALSE} otherwise.
}
\description{
Activate (i.e., set active = TRUE) for a given recipe step.
}
\examples{
\dontrun{

connection <- connect('my_db on localhost')
activate_recipe_step(connection, 42, 256)
disconnect(connection)
}

}
\seealso{
\code{\link[=get_recipe_steps]{get_recipe_steps()}}, \code{\link[=add_recipe_step]{add_recipe_step()}}
}
