% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_instance.R
\name{add_instance}
\alias{add_instance}
\title{Add a new instance to the database}
\usage{
add_instance(connection, name, instance_owner, description = "")
}
\arguments{
\item{connection}{A connection object, as retrieved from \code{\link[=connect_scrapebot]{connect_scrapebot()}}.}

\item{name}{Character string identifying the instance.}

\item{instance_owner}{The email address of the ScrapeBot user who will be the owner of the new instance as character string. If this one does not exist, it will be created (in this case, a text will be raised).}

\item{description}{Character string to describe the instance even further; default is an empty string.}
}
\value{
The \code{uid} (integer) of the newly created instance or an integer \code{NA} in case of error.
}
\description{
Adds a newly created instance to the database. This does mean launching a server (see \code{\link[=aws_launch_instance]{aws_launch_instance()}}).
}
\examples{
\dontrun{

connection <- connect('my_db on localhost')
add_instance(connection, 'local machine', 'mario@haim.it')
disconnect(connection)
}

}
\seealso{
\code{\link[=get_instances]{get_instances()}}, \code{\link[=aws_launch_instance]{aws_launch_instance()}}
}
