% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_recipe.R
\name{add_recipe}
\alias{add_recipe}
\title{Add a new recipe}
\usage{
add_recipe(
  connection,
  name,
  recipe_owner,
  description = "",
  active = FALSE,
  cookies = FALSE,
  interval = 15
)
}
\arguments{
\item{connection}{A connection object, as retrieved from \code{\link[=connect_scrapebot]{connect_scrapebot()}}.}

\item{name}{Character string identifying the recipe.}

\item{recipe_owner}{The email address of the ScrapeBot user who will be the owner of the new recipe as character string. If this one does not exist, it will be created (in this case, a text will be raised).}

\item{description}{Character string to describe the recipe even further; default is an empty string.}

\item{active}{Boolean to indicate whether the recipe should be active; default is \code{FALSE}.}

\item{cookies}{Boolean to indicate whether the recipe should store cookies; default is \code{FALSE}.}

\item{interval}{Integer to indicate the number of minutes after which a recipe should be run; default is \code{15}.}
}
\value{
The \code{uid} (integer) of the newly created recipe or an integer \code{NA} in case of error.
}
\description{
Adds a newly created recipe to the database. This new recipe is empty (i.e., does not contain steps).
Also, this new recipe f
}
\examples{
\dontrun{

connection <- connect('my_db on localhost')
add_recipe(connection, 'visit my website', 'mario@haim.it')
add_recipe(connection, 'visit my website', 'mario@haim.it', active = T)
disconnect(connection)
}

}
\seealso{
\code{\link[=get_recipes]{get_recipes()}}, \code{\link[=add_recipe_step]{add_recipe_step()}}
}
