% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_recipe_step.R
\name{add_recipe_step}
\alias{add_recipe_step}
\title{Add a new single recipe step}
\usage{
add_recipe_step(
  connection,
  recipe_uid,
  type = "log",
  fixed_value = "",
  random_items = c(),
  use_data_item_instead_of_value = FALSE,
  active = TRUE,
  sort = NA_integer_
)
}
\arguments{
\item{connection}{A connection object, as retrieved from \code{\link[=connect_scrapebot]{connect_scrapebot()}}.}

\item{recipe_uid}{The \code{uid} (integer) of the corresponding recipe.}

\item{type}{The type of what this step should do. Has to be one of the following as character string:
\itemize{
\item click
\item data
\item element_screenshot
\item execute_js
\item find_by_class
\item find_by_css
\item find_by_id
\item find_by_link
\item find_by_link_partial
\item find_by_name
\item find_by_tag
\item find_by_xpath
\item get_attribute
\item get_attributes
\item get_element_count
\item get_htmlsource
\item get_pagetitle
\item get_text
\item get_texts
\item get_value
\item get_values
\item go_back
\item go_forward
\item log
\item navigate
\item pause
\item post_all_data
\item post_previous_step_data
\item random_select
\item screenshot
\item scroll_to
\item sometimes_screenshot
\item submit
\item unset_prior_element
\item write
\item write_slowly
}}

\item{fixed_value}{Possibly required value (depends on type); default is an empty character string.}

\item{random_items}{Character vector to be used as urn to draw random items for \code{fixed_value} from.}

\item{use_data_item_instead_of_value}{Logical indicating whether value should be taken from a previous data item. If \code{TRUE}, \code{fixed_value} needs to be set accordingly.}

\item{active}{Boolean to indicate whether the recipe step should be active; default is \code{TRUE}.}

\item{sort}{Numeric index to have steps be run one after another. If \code{NA}, new step is appended to the end.}
}
\value{
The \code{uid} (integer) of the newly created recipe step or an integer \code{NA} in case of error.
}
\description{
Adds a newly created recipe step to the database.
See https://github.com/MarHai/ScrapeBot/blob/master/scrapebot/emulate.py for type descriptions.
}
\examples{
\dontrun{

connection <- connect('my_db on localhost')
recipe_uid <- add_recipe(connection, 'visit my website')
add_recipe_step(connection, recipe_uid, 'navigate', 'https://www.google.com/')
add_recipe_step(connection, recipe_uid, 'screenshot')
disconnect(connection)
}

}
\seealso{
\code{\link[=add_recipe]{add_recipe()}}, \code{\link[=get_recipes]{get_recipes()}}, \code{\link[=get_recipe_steps]{get_recipe_steps()}}
}
