% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{aws_default_ec2_image}
\alias{aws_default_ec2_image}
\title{Get a default image name for a given region}
\usage{
aws_default_ec2_image(region = "eu-central-1")
}
\arguments{
\item{region}{The AWS region to fetch the image name for (e.g., "eu-central-1" for Frankfurt/Europe) as character.}
}
\value{
AWS Amazon Machine Image (AMI) name as character string or (character) \code{NA} if region unknown.
}
\description{
AWS works in regions. And just like instances, also Amazon Machine Images are customized per region (e.g., timezone, language ...).
This function serves as helping dictionary in that it provides you with the AMI image name for a given image.
The image returned here is the one for an Ubuntu Server 20.04 LTS with SSD volume (64-bit x86, if available).
This has proven to work well with both AWS and ScrapeBot.
It is also eligible for the free tier.
}
\examples{

aws_default_ec2_image()
aws_default_ec2_image('eu-central-1')

}
\seealso{
\code{\link[=connect_aws]{connect_aws()}}, \code{\link[=aws_launch_instance]{aws_launch_instance()}}
}
