% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{aws_deploy_ssh_to_ec2}
\alias{aws_deploy_ssh_to_ec2}
\title{Deploy a public SSH key to an AWS EC2 instance.}
\usage{
aws_deploy_ssh_to_ec2(aws_connection)
}
\arguments{
\item{aws_connection}{AWS connection object, as retrieved from \code{\link[=connect_aws]{connect_aws()}}.}
}
\description{
Checks if a keypair with a name compiled from the credential_section and the prefix "ScrapeBotR" (e.g., "ScrapeBotR AWS") exists.
If not, creates a new keypair and imports the public SSH key as provided in \code{\link[=write_aws_credentials]{write_aws_credentials()}}.
}
\examples{
\dontrun{

aws_deploy_ssh_to_ec2(aws_connection)
}

}
\seealso{
\code{\link[=connect_aws]{connect_aws()}}
}
