% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{aws_launch_database}
\alias{aws_launch_database}
\title{Launch a new database as AWS RDS database instance}
\usage{
aws_launch_database(aws_connection, rds_type = "db.m6g.xlarge")
}
\arguments{
\item{aws_connection}{AWS connection object, as retrieved from \code{\link[=connect_aws]{connect_aws()}}. This also specifies the region.}

\item{rds_type}{Just like EC2 instances, also RDS instances require a type. Note that these are not free. The default here, db.m6g.xlarge (along with the here pre-set 50GByte of storage), amounts to roughly 270 USD per month.}
}
\value{
The updated aws_connection object with rds_credential_section (to be used with connect_scrapebot) and rds_identifier (to be used with aws_terminate_database) set.
}
\description{
Launches an RDS instance and installs a fresh ScrapeBot central database.
}
\details{
Note that this may cause costs as RDS instances cost real money.
Also, note that as this function waits for a new instance to launch, it takes quiet some time (usually some 3min).

The RDS instance gets created in the current region with TCP traffic via port 3306 (default MySQL port) open to the world.
Currently uses MySQL 8.0.21 on a single-AZ publicly accessible RDS instance without automated backups and 50 GByte of storage.
The function waits for the instance to boot (this takes some time, up to 3-5min) before calling the \code{\link[=write_scrapebot_credentials]{write_scrapebot_credentials()}} function to store the respective credentials.
Ultimately, the function installs an empty ScrapeBot database, updates the aws_connection object and returns the latter.
}
\examples{
\dontrun{

aws_connection <- connect_aws()
aws_connection <- aws_launch_database(aws_connection)
aws_connection <- aws_launch_database(aws_connection, 'db.m6g.x2large')
}

}
\seealso{
\code{\link[=connect_aws]{connect_aws()}}, \code{\link[=aws_terminate_database]{aws_terminate_database()}}, \code{\link[=get_or_create_user]{get_or_create_user()}}
}
