% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{aws_launch_storage}
\alias{aws_launch_storage}
\title{Launch a new storage as AWS S3 bucket}
\usage{
aws_launch_storage(aws_connection)
}
\arguments{
\item{aws_connection}{AWS connection object, as retrieved from \code{\link[=connect_aws]{connect_aws()}}. This also specifies the region.}
}
\value{
The updated AWS connection object with s3_bucket set (keep/store this to later terminate the bucket also).
}
\description{
The bucket gets created in the current region without particular security settings.
Its name is an automatically generated random character string.
}
\examples{
\dontrun{

aws_connection <- connect_aws()
aws_launch_storage(aws_connection)
}

}
\seealso{
\code{\link[=connect_aws]{connect_aws()}}, \code{\link[=aws_terminate_storage]{aws_terminate_storage()}}
}
