% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{aws_load_connection}
\alias{aws_load_connection}
\title{Load a previously saved AWS connection object.}
\usage{
aws_load_connection(file = "aws_connection.rds")
}
\arguments{
\item{file}{Path to the file where to save the aws_connection object (in RDS format).}
}
\value{
The loaded and re-initiated AWS connection object.
}
\description{
Loads a previously saved AWS connection object.
Then, the function re-initiates the connection to AWS.
}
\examples{
\dontrun{

aws_load_connection('aws_connection.rds')
}

}
\seealso{
\code{\link[=connect_aws]{connect_aws()}}, \code{\link[=aws_load_connection]{aws_load_connection()}}
}
