% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{aws_save_connection}
\alias{aws_save_connection}
\title{Save an AWS connection object for later use.}
\usage{
aws_save_connection(aws_connection, file = "aws_connection.rds")
}
\arguments{
\item{aws_connection}{AWS connection object, as retrieved from \code{\link[=connect_aws]{connect_aws()}}.}

\item{file}{Path to the file where to save the aws_connection object (in RDS format).}
}
\value{
The unchanged AWS connection object for chaining.
}
\description{
AWS connection objects hold information about running instances, storages, and set up databases.
To later terminate them, you need to store/save your AWS connection object.
}
\examples{
\dontrun{

aws_save_connection(aws_connection)
}

}
\seealso{
\code{\link[=connect_aws]{connect_aws()}}, \code{\link[=aws_load_connection]{aws_load_connection()}}
}
