% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{aws_terminate_database}
\alias{aws_terminate_database}
\title{Terminate an AWS RDS database instance}
\usage{
aws_terminate_database(aws_connection)
}
\arguments{
\item{aws_connection}{AWS connection object, as retrieved from \code{\link[=connect_aws]{connect_aws()}}. This also specifies the region.}
}
\value{
The updated AWS connection object.
}
\description{
Terminating an RDS database instance means deleting it and all its contents, backups, and images.
The RDS database instance to be deleted is taken from the aws_connection object (\verb{$rds_identifier}) which itself is set in \code{\link[=aws_launch_database]{aws_launch_database()}}.
This is major, do that with great care!
}
\examples{
\dontrun{

aws_terminate_database(aws_connection)
}

}
\seealso{
\code{\link[=connect_aws]{connect_aws()}}, \code{\link[=aws_launch_storage]{aws_launch_storage()}}
}
