% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{aws_terminate_instance}
\alias{aws_terminate_instance}
\title{Terminate an AWS EC2 instance}
\usage{
aws_terminate_instance(aws_connection, instance_aws_uid)
}
\arguments{
\item{aws_connection}{AWS connection object, as retrieved from \code{\link[=connect_aws]{connect_aws()}}. This also specifies the region.}

\item{instance_aws_uid}{Character string with the instance to be terminated. This has been put into the aws_connection$ec2_instances tibble.}
}
\value{
The updated AWS connection object.
}
\description{
Terminating an EC2 instance translates to shutting down and deleting it. That means, all data stored on that machine is gone.
This does not affect a central database, given that the central database is located on a different machine.
Note that AWS lists an instance after shutting it down for some 10-20 minutes longer. You won't be charged for such instances, though.
}
\examples{
\dontrun{

aws_connection <- connect_aws()
aws_terminate_instance(aws_connection, 'i-0a1b2c3d4e5f6g7h8i9j')
}

}
\seealso{
\code{\link[=connect_aws]{connect_aws()}}, \code{\link[=aws_launch_instance]{aws_launch_instance()}}
}
