% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{aws_terminate_storage}
\alias{aws_terminate_storage}
\title{Terminate an AWS S3 bucket}
\usage{
aws_terminate_storage(aws_connection)
}
\arguments{
\item{aws_connection}{AWS connection object, as retrieved from \code{\link[=connect_aws]{connect_aws()}}. This also specifies the region.}
}
\value{
The updated AWS connection object.
}
\description{
Terminating an S3 bucket means deleting it and all its contents.
The S3 bucket to be deleted is taken from the aws_connection object (\verb{$s3_bucket}) which itself is set in \code{\link[=aws_launch_storage]{aws_launch_storage()}}.
This is major, do that with great care!
}
\examples{
\dontrun{

aws_terminate_storage(aws_connection)
}

}
\seealso{
\code{\link[=connect_aws]{connect_aws()}}, \code{\link[=aws_launch_storage]{aws_launch_storage()}}
}
