% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{change_aws_region}
\alias{change_aws_region}
\title{Change the current region in your Amazon Web Services (AWS)}
\usage{
change_aws_region(aws_connection, new_region = "eu-central-1")
}
\arguments{
\item{aws_connection}{AWS connection object, as retrieved from \code{\link[=connect_aws]{connect_aws()}}. Holds the current region.}

\item{new_region}{The new AWS region to connect to (e.g., "eu-central-1" for Frankfurt) as character.}
}
\value{
The updated AWS connection object.
}
\description{
AWS works in regions. As such, you need to specify the region when connecting for all subsequent actions. You AWS user might not have all regions enabled. Find out more about it in your account settings (https://console.aws.amazon.com/billing/home?#/account).
}
\examples{
\dontrun{

aws_connection <- connect_aws()
change_aws_region(aws_connection, 'us-east-2')
}

}
\seealso{
\code{\link[=connect_aws]{connect_aws()}}
}
