% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Connect to database}
\usage{
connect(host, user, password, database = "scrapebot")
}
\arguments{
\item{host}{The central database host to connect to as character string.}

\item{user}{The username to connect to the central database as character string.}

\item{password}{According password as character string.}

\item{database}{Represents the database name as character string. Defaults to \code{scrapebot}.}
}
\value{
A connection object (i.e., a specified list), ready to be passed to other \link{ScrapeBotR} functions.
}
\description{
Connect to a ScrapeBot central database as typically provided during the installation process of a ScrapeBot infrastructure.
Uses RMariaDB to connect to the MySQL server.
}
\examples{
\dontrun{

connect('localhost', 'root', 's3cr3t_password')
connect('127.0.0.1', 'root', 's3cr3t_password', 'my_db')
}

}
\seealso{
\code{\link[=disconnect]{disconnect()}} to close a database connection, \url{https://github.com/MarHai/ScrapeBot} for more information on the central database.
}
