% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect_scrapebot}
\alias{connect_scrapebot}
\title{Connect to ScrapeBot database}
\usage{
connect_scrapebot(credentials_section, credentials_file = "~/.scrapebot.ini")
}
\arguments{
\item{credentials_section}{The section within your INI file holding the credentials. If \code{\link[=write_scrapebot_credentials]{write_scrapebot_credentials()}} was used to create the INI file, then the section is called "\code{database} on \code{host}" (e.g., "scrapebot on localhost").}

\item{credentials_file}{The INI file to use. If \code{\link[=write_scrapebot_credentials]{write_scrapebot_credentials()}} was used to create the INI file, then it is called \code{~/.scrapebot.ini} (the default). Note that the file must (!) be located inside your home directory (i.e., somewhere inside \code{~} as this is a required by \url{https://mariadb.com/kb/en/configuring-mariadb-with-option-files/}).}
}
\value{
A connection object (i.e., a specified list), ready to be passed to other \link{ScrapeBotR} functions.
}
\description{
Connect to a ScrapeBot central database as typically provided during the installation process of a ScrapeBot infrastructure.
Uses RMariaDB to connect to the MySQL server.
}
\details{
In order not to use database credentials here, you need to set up an INI file containing your database credentials. You can do this by hand (see \code{\link[=write_scrapebot_credentials]{write_scrapebot_credentials()}}) or use the \code{\link[=write_scrapebot_credentials]{write_scrapebot_credentials()}} function.
}
\examples{
\dontrun{

connect_scrapebot('my_db on localhost')
connect_scrapebot('my_db on localhost', '~/my_own_credentials_file.ini')
}

}
\seealso{
\code{\link[=write_scrapebot_credentials]{write_scrapebot_credentials()}} to help create the credentials file, \code{\link[=disconnect]{disconnect()}} to close a database connection, \url{https://github.com/MarHai/ScrapeBot} for more information on the central database.
}
