% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate_recipe.R
\name{deactivate_recipe}
\alias{deactivate_recipe}
\title{Deactivate (i.e., set active = FALSE) for a given recipe.}
\usage{
deactivate_recipe(connection, recipe_uid)
}
\arguments{
\item{connection}{A connection object, as retrieved from \code{\link[=connect_scrapebot]{connect_scrapebot()}}.}

\item{recipe_uid}{The \code{uid} (integer) of the corresponding recipe.}
}
\value{
\code{TRUE} if update was successful, \code{FALSE} otherwise.
}
\description{
Deactivate (i.e., set active = FALSE) for a given recipe.
}
\examples{
\dontrun{

connection <- connect('my_db on localhost')
deactivate_recipe(connection, 42)
disconnect(connection)
}

}
\seealso{
\code{\link[=get_recipes]{get_recipes()}}, \code{\link[=add_recipe]{add_recipe()}}
}
