% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_sbj.R
\name{export_sbj}
\alias{export_sbj}
\title{Export a ScrapeBot JSON recipe file (\code{.sbj})}
\usage{
export_sbj(connection, recipe_uid, sbj_file = NULL)
}
\arguments{
\item{connection}{A connection object, as retrieved from \code{\link[=connect_scrapebot]{connect_scrapebot()}}.}

\item{recipe_uid}{The recipe's \code{UID}, for which the SBJ should be exported.}

\item{sbj_file}{If set to a file path, the SBJ file is directly written to this path. Otherwise (or \code{NULL}), character-string representation is returned.}
}
\value{
If \code{sbj_file} is set to a valid file path, this path is returned after a successful write. Otherwise, the SBJ raw text is returned.
}
\description{
Given a recipe (\code{UID}), this function exports the recipe's SBJ representation.
}
\examples{
\dontrun{

connection <- connect('my_db on localhost')
export_sbj(connection, 42)
export_sbj(connection, 42, 'my_recipe.sbj')
disconnect(connection)
}

}
\seealso{
\code{\link[=import_sbj]{import_sbj()}}
}
