% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instances.R
\name{get_instances}
\alias{get_instances}
\title{Get a tibble of all available instances}
\usage{
get_instances(connection)
}
\arguments{
\item{connection}{A connection object, as retrieved from \code{\link[=connect]{connect()}}.}
}
\value{
A \link[tibble:tibble-package]{tibble} listing all available instances including its UID, name, the date then the instance was first created, its description, as well as the number of runs on that instance and the date of the latest run on that instance.
}
\description{
Get a tibble of all available instances
}
\examples{
\dontrun{

connection <- connect('localhost', 'root', 's3cr3t_password')
get_instances(connection)
disconnect(connection)
}

}
