% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_or_create_user.R
\name{get_or_create_user}
\alias{get_or_create_user}
\title{Get a tibble of all available instances}
\usage{
get_or_create_user(connection, email)
}
\arguments{
\item{connection}{A ScrapeBot connection object, as retrieved from \code{\link[=connect_scrapebot]{connect_scrapebot()}}.}

\item{email}{Character string with the email address to be used for the user.}
}
\value{
A \link[tibble:tibble-package]{tibble} listing all available instances including its UID, name, the date then the instance was first created, its description, as well as the number of runs on that instance and the date of the latest run on that instance.
}
\description{
Get a tibble of all available instances
}
\examples{
\dontrun{

connection <- connect('my_db on localhost')
get_or_create_user(connection)
}

}
