% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipes.R
\name{get_recipes}
\alias{get_recipes}
\title{Get a tibble of all recipes}
\usage{
get_recipes(connection, instance_uid = NULL, include_inactive = FALSE)
}
\arguments{
\item{connection}{a connection object, as retrieved from [connect()]}

\item{include_inactive}{if true, inactive recipes are included along active recipes; defaults to `FALSE`}
}
\value{
a [tibble][tibble::tibble-package] listing all available recipes including its UID, name, the date then the recipe was first created, its description, whether the recipe is active, as well as the number of its runs and the date of the latest run
}
\description{
Get a tibble of all recipes
}
\examples{
\dontrun{
connection <- connect('localhost', 'root', 's3cr3t_password')
get_recipes(connection)
get_recipes(connection, include_inactive = TRUE)
disconnect(connection)
}

}
\seealso{
[get_recipe_steps()]
}
