% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_run_log.R
\name{get_run_log}
\alias{get_run_log}
\title{Retrieve log entries for a provided run}
\usage{
get_run_log(connection, run_uid)
}
\arguments{
\item{connection}{A connection object, as retrieved from \code{\link[=connect]{connect()}}.}

\item{run_uid}{Numeric UID or a vector of numeric UIDs of the run to find log entries for.}
}
\value{
A \link[tibble:tibble-package]{tibble} listing all log entries including a timestamp, its type (one of: \code{info}, \code{warning}, \code{error}), and the message itself.
}
\description{
Retrieve log entries for a provided run
}
\examples{
\dontrun{

connection <- connect('localhost', 'root', 's3cr3t_password')
get_run_log(connection, 13)
get_run_log(connection, c(13, 99))
disconnect(connection)
}

}
\seealso{
\code{\link[=get_runs]{get_runs()}}, \code{\link[=get_run_data]{get_run_data()}}
}
