% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_runs.R
\name{get_runs}
\alias{get_runs}
\title{Get a tibble of all runs for the provided instance/recipe}
\usage{
get_runs(connection, instance_uid = NULL, recipe_uid = NULL)
}
\arguments{
\item{connection}{a connection object, as retrieved from [connect()]}

\item{instance_uid}{numeric UID of the instance to filter runs for. If NULL, `recipe_uid` has to be provided. Defaults to NULL.}

\item{recipe_uid}{numeric UID of the recipe to filter runs for. If NULL, `instance_uid` has to be provided. Defaults to NULL.}
}
\value{
a [tibble][tibble::tibble-package] listing all runs including their UID, the UID of the run's recipe, the UID of the run's instance, the date when the run started, its runtime (in seconds), and the status (one of: success, error, config_error, command_not_found, in_progress)
}
\description{
A run represents one execution of a recipe on a certain instance. For this function to provide meaningful results, you need to ...
- specify only `instance_uid` to list all runs on a specific instance
- specify only `recipe_uid` to list all runs for a specific recipe
- specify both `instance_uid` and `recipe_uid` to list all runs of a specific recipe on a given instance
}
\details{
The function will yield an error if neither one of instance/recipe is provided.

Note that the resulting tibble might be rather large and the request could take some time.
}
\examples{
\dontrun{
connection <- connect('localhost', 'root', 's3cr3t_password')
get_runs(connection, instance_uid = 42, recipe_uid = 21)
get_runs(connection, instance_uid = 42)
get_runs(connection, recipe_uid = 21)
disconnect(connection)
}

}
\seealso{
[get_run_data()], [get_run_log()]
}
