% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe2instance.R
\name{remove_recipe_from_instance}
\alias{remove_recipe_from_instance}
\title{Remove a recipe from a given instance.}
\usage{
remove_recipe_from_instance(connection, recipe_uid, instance_uid)
}
\arguments{
\item{connection}{A connection object, as retrieved from \code{\link[=connect_scrapebot]{connect_scrapebot()}}.}

\item{recipe_uid}{The \code{uid} (integer) of the corresponding recipe.}

\item{instance_uid}{The \code{uid} (integer) of the corresponding instance.}
}
\value{
\code{TRUE} if removal was successful, \code{FALSE} otherwise.
}
\description{
This function only makes a recipe stop running on a given instance.
It does not really delete, neither the recipe, nor the instance, or any previous runs/data.
}
\examples{
\dontrun{

connection <- connect('my_db on localhost')
remove_recipe_from_instance(connection, 42, 2)
disconnect(connection)
}

}
\seealso{
\code{\link[=get_recipes]{get_recipes()}}, \code{\link[=get_instances]{get_instances()}}
}
