% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_credentials.R
\name{write_scrapebot_credentials}
\alias{write_scrapebot_credentials}
\title{Write ScrapeBot database credentials to INI file}
\usage{
write_scrapebot_credentials(
  host,
  user = NULL,
  password = NULL,
  database = "scrapebot",
  port = 0
)
}
\arguments{
\item{host}{The central database host to connect to as character string.}

\item{user}{The username to connect to the central database as character string.}

\item{password}{According password as character string.}

\item{database}{Represents the database name as character string. Defaults to \code{scrapebot}.}

\item{port}{The port number to connect to (only used if greater than 0; not needed for MySQL if 3306). Defaults to \code{0}.}
}
\value{
The file name of the newly created credentials file, if successful.
}
\description{
You should not use connection functionality by regularly publishing credentials in the console. While credentials will end up in the console log, you also risk pushing scripts including credentials to any sustainable repositories (e.g., Git).
Instead, you should prepare an INI file which will then be read/used by the connection functionality.
}
\details{
This function is a helper function to create this INI file for you.
While creating the INI file like so will still end up pushing your credentials to the console log once,
but it will not require you to do that ever again, thus keeping credentials off your repositories.
The function pushes the credentials into an INI-file section called "\code{database} on \code{host}" (e.g., "scrapebot on localhost").

Alternatively, you can absolutely create this INI file yourself. For this, include a file called \code{.scrapebot.ini} (yes, including the leading dot) in your working directory.
This file then needs a section which is indicated in its own line, looking like \code{[my section]}.
Below this line (and thus, within this section) you can specify the following parameters to connect to your database:
\itemize{
\item host (for example, \code{host=localhost})
\item port (for example, \code{port=3307})
\item user (for example, \code{user=my_personal_user})
\item password (for example, \code{password=abcd3.45d:cba!})
\item database (for example, \code{database=scrapebot})
}
}
\examples{
write_scrapebot_credentials('localhost', 'root', 'my_password')
write_scrapebot_credentials(':memory:')

}
\seealso{
\code{\link[=connect_scrapebot]{connect_scrapebot()}} to use the credential file in practice
}
