% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Connect to database}
\usage{
connect(host, user, password, database = "scrapebot")
}
\arguments{
\item{host}{the central database host to connect to as character string}

\item{user}{the username to connect to the central database as character string}

\item{password}{according password as character string}

\item{database}{represents the database name as character string. Defaults to `scrapebot`.}
}
\value{
a connection object (i.e., a list), ready to be passed to other [ScrapeBotR] functions
}
\description{
Connect to a ScrapeBot central database as typically provided during the installation process of a ScrapeBot infrastructure.
Uses RMariaDB to connect to the MySQL server.
}
\examples{
\dontrun{
connect('localhost', 'root', 's3cr3t_password')
connect('127.0.0.1', 'root', 's3cr3t_password', 'my_db')
}

}
\seealso{
[disconnect()] to close a database connection, \url{https://github.com/MarHai/ScrapeBot} for more information on the central database
}
