% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disconnect.R
\name{disconnect}
\alias{disconnect}
\title{Disconnect current database}
\usage{
disconnect(connection)
}
\arguments{
\item{connection}{A connection object, as retrieved from \code{\link[=connect]{connect()}}.}
}
\description{
Properly closes a database connection. Use this if you are employing multiple connections and/or if you are annoyed by the warnings saying to \verb{call dbDisconnect() when finished working with a connection}.
}
\examples{
\dontrun{

connection <- connect('my_db on localhost')
disconnect(connection)
}

}
\seealso{
\code{\link[=connect]{connect()}}
}
