% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_or_create_user.R
\name{get_or_create_user}
\alias{get_or_create_user}
\title{Based on a given ScrapeBot user's email address, get this user's \code{UID} (or create it first)}
\usage{
get_or_create_user(connection, email)
}
\arguments{
\item{connection}{A ScrapeBot connection object, as retrieved from \code{\link[=connect_scrapebot]{connect_scrapebot()}}.}

\item{email}{Character string with the email address to be used for the user.}
}
\value{
The \code{UID} of the (newly created or found) user (as integer). If a user is created, a text will be shown with the generated password.
}
\description{
Based on a given ScrapeBot user's email address, get this user's \code{UID} (or create it first)
}
\examples{
\dontrun{

connection <- connect('my_db on localhost')
get_or_create_user(connection)
}

}
