% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_run_data.R
\name{get_run_data}
\alias{get_run_data}
\title{Collect (large amounts of) data for a given combination of criteria}
\usage{
get_run_data(
  connection,
  run_uid = NULL,
  instance_uid = NULL,
  recipe_uid = NULL,
  step_uid = NULL
)
}
\arguments{
\item{connection}{a connection object, as retrieved from [connect()]}

\item{run_uid}{optional numeric UID of a specific run to collect data from; if NULL, either `instance_uid` or `recipe_uid` (or both) has to be provided. Defaults to `NULL`.}

\item{instance_uid}{optional numeric UID of the instance to filter data for; if NULL, either `run_uid` or `recipe_uid` (or both) has to be provided. Defaults to `NULL`.}

\item{recipe_uid}{optional numeric UID of the recipe to filter data for; if NULL, either `instance_uid` or `run_uid` (or both) has to be provided. Defaults to `NULL`.}

\item{step_uid}{optional numeric UID of a step to filter data for. Defaults to `NULL`.}
}
\value{
a [tibble][tibble::tibble-package] listing all data including their creation timestamp, the UIDs of its run, instance, recipe, and step, as well as the data value itself
}
\description{
Data is stored heavily within ScrapeBot. Hence, filtering needs to be applied to collect data. Two types of filter are possible:
- Filter on the runs for which to collect data: Specify one or many of run_uid, instance_uid, and recipe_uid. At least one is required.
- Filter on the step for which to collect data: Optionally specify a step_uid.
}
\details{
In general, this function only returns data from "success"-ful runs (see [get_runs()]). It then collects runs matching the first set of criteria (i.e., run_uid, instance_uid, and recipe_uid). Ultimately, it collects data, thereby focusing on a specific step if specified.
}
\examples{
\dontrun{
connection <- connect('localhost', 'root', 's3cr3t_password')
get_run_data(connection, instance_uid = 42)
get_run_data(connection, run_uid = 256)
get_run_data(connection, recipe_uid = 21)
get_run_data(connection, instance_uid = 42, recipe_uid = 21)
get_run_data(connection, instance_uid = 42, recipe_uid = 21, step_uid = 8)
disconnect(connection)
}

}
\seealso{
[get_runs()], [get_run_log()]
}
