% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_run_log.R
\name{get_run_log}
\alias{get_run_log}
\title{Retrieve log entries for a provided run}
\usage{
get_run_log(connection, run_uid)
}
\arguments{
\item{connection}{a connection object, as retrieved from [connect()]}

\item{run_uid}{numeric UID of the run to find log entries for}
}
\value{
a [tibble][tibble::tibble-package] listing all log entries including a timestamp, its type (one of: info, warning, error), and the message itself
}
\description{
Retrieve log entries for a provided run
}
\examples{
\dontrun{
connection <- connect('localhost', 'root', 's3cr3t_password')
get_run_log(connection, 13)
disconnect(connection)
}

}
\seealso{
[get_runs()], [get_run_data()]
}
